# ripcity-backend/rip-city-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rip City Ticket Dispatch API
  description: |
    **The Portland-focused ticket deal monitoring and alerting platform**
    
    Rip City Ticket Dispatch provides real-time ticket deal discovery for Portland sports and music events. Our API powers intelligent deal scoring, venue-specific monitoring, and automated alerting systems.
    
    ## Key Features
    - **Real-time deal monitoring** across major ticket platforms
    - **AI-powered deal scoring** with confidence ratings
    - **Portland venue expertise** including Moda Center, Providence Park
    - **Trail Blazers focused** with special Rip City deals
    - **Production-ready** with rate limiting and error handling
    
    ## Authentication
    Contact us for API key access. Currently in private beta.
    
    ## Rate Limits
    - **Free tier**: 1,000 requests/day
    - **Pro tier**: 10,000 requests/day
    - **Enterprise**: Unlimited
    
    Built with ❤️ for Rip City • Go Blazers! 🏀
  version: 1.0.0
  contact:
    name: Rip City Ticket Dispatch
    url: https://ripcityticketdispatch.works
    email: api@ripcityticketdispatch.works
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://ripcityticketdispatch.works/terms

servers:
  - url: https://api.ripcityticketdispatch.works/v1
    description: Production API
  - url: https://staging-api.ripcityticketdispatch.works/v1
    description: Staging API
  - url: http://localhost:3001/v1
    description: Development API

security:
  - ApiKeyAuth: []

paths:
  /deals:
    get:
      summary: Get Ticket Deals
      description: |
        Retrieve ticket deals for Portland events with intelligent scoring and filtering.
        
        Returns deals sorted by score (hottest deals first) with metadata about savings, confidence, and event details.
      operationId: getDeals
      tags:
        - Deals
      parameters:
        - name: category
          in: query
          description: Filter by event category
          schema:
            type: string
            enum: [sports, music, theater, family, other]
          example: sports
        - name: venue
          in: query
          description: Filter by venue name (partial match)
          schema:
            type: string
          example: "Moda Center"
        - name: maxPrice
          in: query
          description: Maximum ticket price filter
          schema:
            type: integer
            minimum: 0
          example: 200
        - name: minSavings
          in: query
          description: Minimum savings percentage required
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 15
          example: 20
        - name: sortBy
          in: query
          description: Sort order for results
          schema:
            type: string
            enum: [score, savings, price, date]
            default: score
        - name: limit
          in: query
          description: Maximum number of deals to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 10
      responses:
        '200':
          description: Successful response with deals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealsResponse'
              examples:
                blazers_deals:
                  summary: Trail Blazers Game Deals
                  value:
                    deals:
                      - id: "blazers-lakers-2025-06-15"
                        eventId: "vvG1zZpkgCJyAE"
                        eventName: "Portland Trail Blazers vs Los Angeles Lakers"
                        venue: "Moda Center"
                        date: "2025-06-15"
                        time: "19:00"
                        originalPrice: 150
                        currentPrice: 120
                        savings: 30
                        savingsPercentage: 20
                        dealScore: 92
                        confidence: 85
                        platform: "ticketmaster"
                        category: "sports"
                        alertLevel: "hot"
                        tags: ["RIP CITY", "GREAT DEAL", "MODA CENTER"]
                        url: "https://www.ticketmaster.com/event/..."
                    total: 1
                    fetched: 1
                    lastUpdated: "2025-06-10T18:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimit'
        '500':
          $ref: '#/components/responses/InternalError'

  /deals/blazers:
    get:
      summary: Get Trail Blazers Deals
      description: |
        Specialized endpoint for Portland Trail Blazers ticket deals.
        
        Returns deals specifically for Blazers games with Rip City branding and special metadata.
      operationId: getBlazersDeals
      tags:
        - Deals
        - Trail Blazers
      responses:
        '200':
          description: Trail Blazers deals with special metadata
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DealsResponse'
                  - type: object
                    properties:
                      specialMessage:
                        type: string
                        description: Special message for Blazers fans
                        example: "🔥 5 hot Blazers deals available - Rip City!"
              examples:
                playoff_game:
                  summary: Playoff Game Deal
                  value:
                    deals:
                      - id: "blazers-playoff-2025"
                        eventName: "Portland Trail Blazers vs Golden State Warriors - Playoff Game 1"
                        venue: "Moda Center"
                        dealScore: 98
                        tags: ["RIP CITY", "PLAYOFFS", "HUGE SAVINGS"]
                        specialNotes: "🏆 PLAYOFF GAME - Don't miss this!"
                    specialMessage: "🏆 Playoff deals - get your tickets now!"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /venues:
    get:
      summary: Get Portland Venues
      description: |
        Retrieve information about Portland event venues with enhanced metadata.
        
        Includes venue details, capacity, teams, and insider tips for Portland locals.
      operationId: getVenues
      tags:
        - Venues
      responses:
        '200':
          description: List of Portland venues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VenuesResponse'
              examples:
                portland_venues:
                  summary: Major Portland Venues
                  value:
                    venues:
                      - id: "KovZpZAJledA"
                        name: "Moda Center"
                        capacity: 19393
                        type: "arena"
                        teams: ["Portland Trail Blazers"]
                        description: "Home of the Portland Trail Blazers"
                        address: "1 Center Ct, Portland, OR 97227"
                        tips: "Arrive early for Trail Blazers games - traffic can be heavy"
                        popularity: 10
                    total: 1
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  schemas:
    Deal:
      type: object
      required:
        - id
        - eventId
        - eventName
        - venue
        - date
        - currentPrice
        - dealScore
        - confidence
        - platform
        - category
        - alertLevel
      properties:
        id:
          type: string
          description: Unique deal identifier
          example: "blazers-lakers-2025-06-15"
        eventId:
          type: string
          description: External event ID from ticket platform
          example: "vvG1zZpkgCJyAE"
        eventName:
          type: string
          description: Full event name
          example: "Portland Trail Blazers vs Los Angeles Lakers"
        venue:
          type: string
          description: Venue name
          example: "Moda Center"
        date:
          type: string
          format: date
          description: Event date
          example: "2025-06-15"
        time:
          type: string
          description: Event time (may be TBA)
          example: "19:00"
        originalPrice:
          type: number
          description: Original/estimated ticket price
          example: 150
        currentPrice:
          type: number
          description: Current ticket price
          example: 120
        savings:
          type: number
          description: Dollar amount saved
          example: 30
        savingsPercentage:
          type: number
          description: Percentage savings
          example: 20
        dealScore:
          type: integer
          minimum: 0
          maximum: 100
          description: AI-calculated deal score (higher = better deal)
          example: 92
        confidence:
          type: integer
          minimum: 0
          maximum: 100
          description: Confidence in deal scoring accuracy
          example: 85
        platform:
          type: string
          enum: [ticketmaster, stubhub, seatgeek, vivid_seats]
          description: Ticket platform source
        section:
          type: string
          description: Seating section
          example: "Lower Bowl"
        row:
          type: string
          description: Seating row
          example: "12"
        seatCount:
          type: integer
          description: Number of seats
          example: 2
        category:
          type: string
          enum: [sports, music, theater, family, other]
          description: Event category
        alertLevel:
          type: string
          enum: [hot, warm, good, normal]
          description: Deal alert priority level
        tags:
          type: array
          items:
            type: string
          description: Descriptive tags for the deal
          example: ["RIP CITY", "GREAT DEAL", "MODA CENTER"]
        url:
          type: string
          format: uri
          description: Link to purchase tickets
          example: "https://www.ticketmaster.com/event/..."
        imageUrl:
          type: string
          format: uri
          description: Event image URL
          example: "https://s1.ticketm.net/dam/a/..."
        lastUpdated:
          type: string
          format: date-time
          description: When deal was last updated
          example: "2025-06-10T18:30:00Z"

    DealsResponse:
      type: object
      required:
        - deals
        - total
        - fetched
        - lastUpdated
      properties:
        deals:
          type: array
          items:
            $ref: '#/components/schemas/Deal'
        total:
          type: integer
          description: Total number of deals found (before limit)
          example: 25
        fetched:
          type: integer
          description: Number of deals returned in this response
          example: 10
        lastUpdated:
          type: string
          format: date-time
          description: When data was last refreshed
          example: "2025-06-10T18:30:00Z"

    Venue:
      type: object
      required:
        - id
        - name
        - popularity
      properties:
        id:
          type: string
          description: Venue ID from ticket platform
          example: "KovZpZAJledA"
        name:
          type: string
          description: Venue name
          example: "Moda Center"
        address:
          type: string
          description: Street address
          example: "1 Center Ct, Portland, OR 97227"
        city:
          type: string
          example: "Portland"
        state:
          type: string
          example: "OR"
        postalCode:
          type: string
          example: "97227"
        capacity:
          type: integer
          description: Venue capacity
          example: 19393
        type:
          type: string
          enum: [arena, stadium, theater, concert_hall, club]
          description: Venue type
        teams:
          type: array
          items:
            type: string
          description: Home teams for this venue
          example: ["Portland Trail Blazers"]
        description:
          type: string
          description: Venue description
          example: "Home of the Portland Trail Blazers"
        parking:
          type: string
          description: Parking information
          example: "Multiple lots and garages available"
        tips:
          type: string
          description: Local insider tips
          example: "Arrive early for Trail Blazers games - traffic can be heavy"
        popularity:
          type: integer
          minimum: 1
          maximum: 10
          description: Popularity score for deal weighting
          example: 10

    VenuesResponse:
      type: object
      required:
        - venues
        - total
      properties:
        venues:
          type: array
          items:
            $ref: '#/components/schemas/Venue'
        total:
          type: integer
          description: Total number of venues
          example: 15
        lastUpdated:
          type: string
          format: date-time
          example: "2025-06-10T18:30:00Z"

    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: "Bad Request"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid category parameter"
        code:
          type: string
          description: Error code for debugging
          example: "INVALID_CATEGORY"

  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Bad Request"
            message: "Invalid category parameter. Must be one of: sports, music, theater, family, other"
            code: "INVALID_CATEGORY"

    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate Limit Exceeded"
            message: "API rate limit exceeded. Try again in 60 seconds."
            code: "RATE_LIMIT_EXCEEDED"
      headers:
        X-RateLimit-Limit:
          description: Requests per hour allowed
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Requests remaining in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when rate limit resets
          schema:
            type: integer

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred while processing your request"
            code: "INTERNAL_ERROR"

tags:
  - name: Deals
    description: Ticket deal discovery and monitoring
  - name: Trail Blazers
    description: Portland Trail Blazers specific endpoints
  - name: Venues
    description: Portland venue information and metadata